ref(py3): Fix buffers to coerce redis values from bytes when needed
This essentially implements the "unicode sandwich" for the redis usage
in the redis buffers backend.

Redis will encode keys and values to utf-8 automatically upon any set
calls, but will not decode for us. This is important in this case
because we do store some pickles in buffer keys, which if decoded would
break.

I've handled decoding where we need to

 1. All hget keys are coerced to strings, these will always be strings.
 2. Values are coerced for json decoding just before decoding
 3. In tests I decode keys for ease of readability